+++ src/views/Home/Home.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import MyCalendar from '../../components/Calendar';
-import UserOverview from '../UserOverview/userOverview';
+import UserOverview from '../../components/UserOverview/userOverview';
 
 export default function HomeRoot() {
   return (
